Implement Safety Legend Toggle
This commit contains the work to create a safety legend toggle for the building view of the application.

This involved connecting an event listener to the toggle switch in the dropdown menu. This event listener updates a value on the site’s cookie (so the user’s selection is remembered between visits). After setting this toggle, the display style of all associated HTML elements are displayed by modifying their display CSS attribute.